requirements.txt
PyQt6>=6.4.0
openai-whisper>=20231117
